pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ML-expert"
version = "0.1.0"
description = "ML research paper processing and dataset creation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "torch>=2.5.1",
    "transformers>=4.46.2",
    "sglang>=0.1.11",  # Added for olmocr
    "accelerate>=0.27.1",  # Added for model loading

    # PDF Processing
    "pypdf>=5.2.0",
    "pypdfium2",
    "pdf2image",  # Added for PDF rendering
    "poppler-utils",  # Added for PDF processing

    # Text Processing
    "lingua-language-detector",
    "ftfy",
    "bleach",
    "markdown2",

    # Image Processing
    "Pillow",

    # File handling and data formats
    "smart_open",
    "zstandard",
    "orjson",
    "cached-path",

    # Cloud and API
    "boto3",
    "httpx",
    "requests",
    "cryptography",

    # Utilities
    "tqdm",
    "filelock",
    "asyncio",

    # olmocr specific
    "olmocr @ git+https://github.com/allenai/olmocr.git",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "black",
    "isort",
    "pytest",
    "pytest-asyncio",
]

[tool.black]
line-length = 160
include = '\.pyi?$'
exclude = '''
(
      __pycache__
    | \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.vscode
    | \.venv
    | \bdist\b
    | \bdoc\b
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.ruff]
line-length = 160
target-version = "py311"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = "tests/"
python_classes = ["Test*", "*Test"]
log_format = "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
log_level = "DEBUG"
asyncio_mode = "auto"

# [project]
# name = "ml-expert"
# version = "0.1.0"
# description = "Add your description here"
# readme = "README.md"
# requires-python = ">=3.11"
# dependencies = [
#     "accelerate>=1.4.0",
#     "arxiv>=2.1.3",
#     "beaker-py>=1.34.1",
#     "black>=25.1.0",
#     "bleach>=6.2.0",
#     "boto3>=1.37.4",
#     "cached-path>=1.6.7",
#     "cryptography>=44.0.2",
#     "datasets>=3.3.2",
#     "einops>=0.8.1",
#     "filelock>=3.17.0",
#     "ftfy>=6.3.1",
#     "httpx>=0.28.1",
#     "isort>=6.0.1",
#     "lingua-language-detector>=2.0.2",
#     "markdown2>=2.5.3",
#     "mypy>=1.15.0",
#     "olmocr",
#     "omegaconf>=2.3.0",
#     "orjson>=3.10.15",
#     "pdftotext>=3.0.0",
#     "peft>=0.14.0",
#     "pillow>=11.1.0",
#     "pypdf>=5.2.0",
#     "pypdfium2>=4.30.1",
#     "pytest>=8.3.5",
#     "pytorch-lightning>=2.5.0.post0",
#     "requests>=2.32.3",
#     "ruff>=0.9.9",
#     "s3fs>=0.4.2",
#     "smart-open>=7.1.0",
#     "tensorboard>=2.19.0",
#     "torch>=2.5.1",
#     "tqdm>=4.67.1",
#     "transformers>=4.46.2",
#     "wandb>=0.19.7",
#     "zstandard>=0.23.0",
# ]

# [tool.uv.sources]
# olmocr = { git = "https://github.com/allenai/olmocr.git" }
